… test/fixtures/generator/bwb.xcodeproj/project.pbxproj
… test/fixtures/generator/bwb.xcodeproj/rules_xcodeproj/external.xcfilelist
… test/fixtures/generator/bwb.xcodeproj/rules_xcodeproj/swift_debug_settings.py
… test/fixtures/generator/bwb_spec.json
… test/fixtures/generator/bwx.xcodeproj/project.pbxproj
… test/fixtures/generator/bwx.xcodeproj/rules_xcodeproj/external.xcfilelist
… test/fixtures/generator/bwx.xcodeproj/rules_xcodeproj/swift_debug_settings.py
… test/fixtures/generator/bwx_spec.json
… third_party/com_github_michaeleisel_jjliso8601dateformatter/BUILD
… third_party/com_github_michaeleisel_jjliso8601dateformatter/missing_include.patch