Currently results in about a 2.5x speedup. For a specific large real world application this saves over 5 seconds of generation time.

| Before | After |
| --- | --- |
| <img width="1055" alt="JSONDecoder taking 9.27s in Instruments" src="https://user-images.githubusercontent.com/158658/197020050-665acfc6-a75f-4d9c-a398-08459bb46aeb.png"> |  <img width="1055" alt="ZippyJSONDecoder taking 3.6s in Instruments" src="https://user-images.githubusercontent.com/158658/197020334-b0336fa2-2933-410f-b851-ce5a97c79c12.png"> |